refactor(session-store): extract snapshot-merge infrastructure (file #4)
Lift the generic snapshot-merge-by-id machinery (SnapshotWithId, the
operation/interaction index WeakMaps, mergeSnapshotsById, AppendedSnapshotIndexMap,
merged-array Proxy views, mergeOperation/InteractionSnapshots) into snapshot-merge.ts;
session-store re-exports the public surface. Pure canonical-data transforms.
Net 13/13 + store vitest parity (503 pass), check 0.
